Fix format conversions and duplicate declarations
-Wformat-signedness reported 964 printf and scanf directives whose
argument is unsigned. Index typedefs (room, object, mobile, zone and
the other vnum and rnum types) now use PRI_IDX or SCN_IDX, plain
unsigned values use %u. Three object listings printed the object count
through PRI_IDX and now use %d, and WILD_DEBUG_MEM cast only the first
operand of its size argument.

Redundant declarations are resolved toward the header that matches the
defining source file. act.h includes movement/movement.h, handler.h the
mob memory and utility headers, and oasis.h mob/mob_autoroll.h, instead
of each carrying its own copies; source files that relied on a removed
copy include the owning header. Block-scope externs for conn2 and conn3
move to mysql.h, and the remaining nested externs use their headers.

strlcat is probed like strlcpy (HAVE_STRLCAT) so the local fallback no
longer redeclares the glibc 2.38 function.

GCC 16.2: format= falls from 964 to 1, redundant-decls from 567 to 0,
nested-externs from 44 to 0.

+++ src/net/protocol.c
@@ -4189,11 +4189,11 @@ static const char *GetRGBColour(bool_t abBackground, int aRed, int aGreen, int a
   int ColVal = 16 + (aRed * 36) + (aGreen * 6) + aBlue;
   int Written;
 
-  Written =
-      snprintf(Result, sizeof(Result), "\033[%c8;5;%c%c%cm", '3' + abBackground, /* Background */
-               '0' + (ColVal / 100),                                             /* Red        */
-               '0' + ((ColVal % 100) / 10),                                      /* Green      */
-               '0' + (ColVal % 10));                                             /* Blue       */
+  Written = snprintf(Result, sizeof(Result), "\033[%c8;5;%c%c%cm",
+                     '3' + (int)abBackground,     /* Background */
+                     '0' + (ColVal / 100),        /* Red        */
+                     '0' + ((ColVal % 100) / 10), /* Green      */
+                     '0' + (ColVal % 10));        /* Blue       */
   if (Written < 0 || (size_t)Written >= sizeof(Result))
     return s_Clean;
   return Result;
